Fix suivis recalculation with MySqlDAO
MySqlDAO doesn't save suivis. The history must be
recalculated after each loading.
Until now, the recalculation in EcritureController
was called when closing MySqlDAO. But at this stage,
MySqlDAO wasn't yet defined as the new DAOFactory.
For that reason, the recalculation was computed from
the previous DAOFactory instance, and saved (!) into
this previous instance, instead of into MySqlDAO.
Now, DAOFactory launches the recalculation before
ending setFactory method, only in the case the model
needs it.